move models/helper -> helpers/model
1) better separation of concerns
2) even greater ~~arrowiness~~
