+++ app/models/menu.js
@@ -1,4 +1,4 @@
-import ModelHelper from 'models/helper';
+import ModelHelper from 'helpers/model';
 import _ from 'lodash';
 
 var Model = new ModelHelper();
